app/requirements.txt
pymongo==4.6.2
loguru==0.7.1
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
langchain==0.1.13
langchain-openai==0.1.1
transformers==4.39.2
anthropic==0.25.0
mangum
transformers[torch]
python-multipart==0.0.9
SpeechRecognition==3.10.3
pydub==0.25.1
joblib==1.4.0
retry==0.9.2
deepmultilingualpunctuation==1.0.1